requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
lxml==6.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==70.2.0
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0+cpu
tqdm==4.67.1
transformers==4.55.4
typing-inspection==0.4.1
typing_extensions==4.14.1
Unidecode==1.4.0
urllib3==2.5.0
